Revert "fix: ensure the vendor navigation listener is set before resolving he configure method promise"
